+++ src/virtual_client.h
@@ -11,6 +11,8 @@
 #include <boost/asio/deadline_timer.hpp>
 
 // WebRTC
+#include <api/media_stream_interface.h>
+#include <api/scoped_refptr.h>
 #include <api/stats/rtc_stats_collector_callback.h>
 #include <api/stats/rtc_stats_report.h>
 
@@ -30,7 +32,7 @@ struct VirtualClientStats {
 };
 
 struct VirtualClientConfig {
-  rtc::scoped_refptr<webrtc::VideoTrackSourceInterface> capturer;
+  webrtc::scoped_refptr<webrtc::VideoTrackSourceInterface> capturer;
   sora::SoraSignalingConfig sora_config;
 
   int max_retry = 0;
@@ -90,10 +92,10 @@ class VirtualClient : public std::enable_shared_from_this<VirtualClient>,
   void OnPush(std::string text) override {}
   void OnMessage(std::string label, std::string data) override {}
 
-  void OnTrack(rtc::scoped_refptr<webrtc::RtpTransceiverInterface> transceiver)
+  void OnTrack(webrtc::scoped_refptr<webrtc::RtpTransceiverInterface> transceiver)
       override {}
   void OnRemoveTrack(
-      rtc::scoped_refptr<webrtc::RtpReceiverInterface> receiver) override {}
+      webrtc::scoped_refptr<webrtc::RtpReceiverInterface> receiver) override {}
 
   void OnDataChannel(std::string label) override {}
 
@@ -107,8 +109,8 @@ class VirtualClient : public std::enable_shared_from_this<VirtualClient>,
   std::function<void(std::string)> on_close_;
   boost::asio::deadline_timer retry_timer_;
   std::shared_ptr<sora::SoraSignaling> signaling_;
-  rtc::scoped_refptr<webrtc::AudioTrackInterface> audio_track_;
-  rtc::scoped_refptr<webrtc::VideoTrackInterface> video_track_;
+  webrtc::scoped_refptr<webrtc::AudioTrackInterface> audio_track_;
+  webrtc::scoped_refptr<webrtc::VideoTrackInterface> video_track_;
   
   // OnSetOffer で取得した情報を保存
   mutable std::mutex stats_mutex_;
@@ -133,10 +135,12 @@ class StatsCollectorCallback : public webrtc::RTCStatsCollectorCallback {
       : client_(client) {}
 
   void OnStatsDelivered(
-      const rtc::scoped_refptr<const webrtc::RTCStatsReport>& report) override;
+      const webrtc::scoped_refptr<const webrtc::RTCStatsReport>& report) override;
 
  private:
   std::weak_ptr<VirtualClient> client_;
+  webrtc::scoped_refptr<webrtc::AudioTrackInterface> audio_track_;
+  webrtc::scoped_refptr<webrtc::VideoTrackInterface> video_track_;
 };
 
 #endif